move simplespawner into jupyterhub.spawner
